% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/at_pm.R
\docType{data}
\name{at_pm}
\alias{at_pm}
\title{Particulate Matter (PM)}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 141722 rows and 7 columns.}
\usage{
at_pm
}
\description{
This dataset contains the measures of PM10 and PM2.5 for the years 2008-2014.
}
\author{
Patrick Hausmann, Source: ARPAT Toscana
}
\keyword{datasets}

